Our player-attack function needs two helper functions
The first one it needs is a random-monster function that picks a monster
to target for the chaotic roundhouse attack, while ensuring that the
chosen monster isn't already dead.

The player-attack function also needs a function that allows the player
to pick a monster to target for the nonrandom attacks. This is the job
of the pick-monster function.

Next part is Monster Management Functions on page 197.