The Generic Monster - monster structure
Different monster have one thing in common: a health meter that
determines how many hits they can take before they die.

This is captured in a monster structure. The default value for health
slot is evaluated when a new monster is created. Since this form calls
randval function, every monster will start the battle with a different,
random, health.

To create (and test it) a monster execute: (make-monster).