The Generic Monster - monster-hit function
The function is needed which takes away a monster's health when it's
attacked. It displays a message explaining what happened, and showing if
the monster has just died.

The function used here is not defined with defun, but with defmethod,
which  will display special messages when the knight beats on particular
monsters.