The Wicked Orc
The orc is a simple foe. He can deliver a strong attack with his club,
but otherwise he is pretty harmless. Every orc has a club with a unique
attack level. Orcs are best ignored, unless there are orcs with an
unusually powerful club attack that you want to cull from the herd at
the beginning of a battle.

To create the orc, we define an orc datatype witch includes all the
fields of monster, such as health.

There are two functions for orcs: monster-show and monster-attack.

The one unique thing about orc type is that each orc has an
orc-club-level field and it is used in the two functions.