The Malicious Hydra
The hydra is a very nasty enemy. It will attack you with its many heads,
which you'll need to chop off to defeat it. The hydra's special power is
that it can grow a new head during each round of battle, which means you
want to defeat it as early as possible.

The code for handling the hydra is similar to the code for handling the
orc. The main difference is that a hydra's health also acts as a
stand-in for the number of hydra heads.

A hydra also does something interesting when it is hit: it loses heads.
Thus, there is a special function monster-hit.

Another interesting feature of hydra is that it increments the health
with every attack, so that the hydra grows a new head every turn.