The Slimy Slime Mold
This monster is unique. When it attacks you, it will wrap itself around
your legs and immobilize you, letting the other bad guys finish you off.
It can also squirt goo in your face. You must think quickly in battle to
decide if it's better to finish the slime off early in order to maintain
your agility, or ignore it to focus on more vicious foes first. The
thing is that when your agility is decreased, your ability to attack is
decreased.